## Unnest LibraryExpressionsBuilder and its context from ExpressionBuilder. Library removed from ExpressionBuilder, instead access via ExpressionBuilderContext.LibraryContext.Library
